+++ implementations/meth1/lib/util.hh
@@ -23,9 +23,9 @@ const std::string join( const std::vector<std::string> & command );
 template <typename T> inline void zero( T & x ) { memset( &x, 0, sizeof( x ) ); }
 
 /* Physical memory present on the machine */
-uint64_t memory_exists( void );
+size_t memory_exists( void );
 
 /* Physical memory present and free on the machine */
-uint64_t memory_free( void );
+size_t memory_free( void );
 
 #endif /* UTIL_HH */
